backend/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
Django==4.1.7
django-extensions==3.2.1
django-graphql-jwt==0.3.4
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
Pillow==9.4.0
promise==2.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
